added bypass (#18)
added a runtime check for josepy version compatibility, along with a env-var to bypass.

josepy is not a requirement, so is unpinned as a core dependency; it is pinned as an extras (`pip install cert_utils[josepy]`) – but there exists situations where a potentially-incompatible version can get pulled in.  To avoid this situation, the version of josepy is checked during runtime.  on py3.7 this requires `importlib_metadata`, as that package was not merged into the stdlib until 3.8